Sullivanplot allows log scale, more control
Added an option to do a log scale on the yaxis -- or to pass any valid string to ax.set_yscale(). Also shifted code to set xlabel and ylabel to the case where there's not a passed axis, otherwise you have to hand-code. This gives more flexibility when making multi-subplot figures