added caching for S
Added a mechanism to cache S values once calculated. Since clique calculation is the slowest point down, it's better not to have to recalc this, and we don't need to with a stable graph. Note that right now calculating pi calls S, so for naive looping this gets done twice.  If paths are cleared so is this cache. passes unit tests.